cleaned up some canvases and worked on time handler

+++ Assets/Scripts/ChildPower.cs
@@ -6,14 +6,14 @@ public class ChildPower {
 	//Need power name, mother power, father power, point values for str, agility, will, and fortitude, need base power popularity value, etc.
 	public string powerName { get; set; } 
   public string description { get; set; }
-	public float Strength { get; set; } //0 - 100
-	public float Agility { get; set; } //0 - 100
-	public float Will { get; set; }  //0 - 100
-	public float Fortitude { get; set; } //0 - 100
-	public float Popularity { get; set; } //0 - 100
+	public double Strength { get; set; } //0 - 100
+	public double Agility { get; set; } //0 - 100
+	public double Will { get; set; }  //0 - 100
+	public double Fortitude { get; set; } //0 - 100
+	public double Popularity { get; set; } //0 - 100
 
 	public ChildPower (ParentalPower fPower, ParentalPower mPower) {
-    float[] childAttributes = getGenerationalAttributes(fPower, mPower);
+    double[] childAttributes = getGenerationalAttributes(fPower, mPower);
     string fPowerName = fPower.PowerName;
     string mPowerName = mPower.PowerName;
     Strength = childAttributes[0];
@@ -21,10 +21,11 @@ public class ChildPower {
     Will = childAttributes[2];
     Fortitude = childAttributes[3];
     //We're gonna get into a lot of stuff here alright?
+    //TODO: SEPARATE OUT INTO ITS OWN JSON FILE
     if (fPowerName == "SUPER STRENGTH" && mPowerName == "SUPER STRENGTH") {
       powerName = "DEVESTATING MIGHT";
       description = "Your muscles crackle with power. Even at such a young age your muscles bulge, barely able to contain the raw might stored within them. You've had very little run-ins with villains up to this point, and those you've faced buckled under your sheer power. Your enemies know best not to cross you.";
-      Strength = (int)(this.Strength * 1.4f);
+      Strength = this.Strength * 1.4f;
     }
     if ((fPowerName == "SUPER STRENGTH" && mPowerName == "TELEKINESIS") || (fPowerName == "TELEKINESIS" && mPowerName == "SUPER STRENGTH")) {
       powerName = "CONQUER";
@@ -37,12 +38,12 @@ public class ChildPower {
     }
     if ((fPowerName == "SUPER STRENGTH" && mPowerName == "SHAPESHIFT") || (fPowerName == "SHAPESHIFT" && mPowerName == "SUPER STRENGTH")) {
       powerName = "RAVAGE FORM";
-      Fortitude = (int)(this.Fortitude * 1.2f);
+      Fortitude = this.Fortitude * 1.2f;
       description = "Your muscles bulge and tear, your bones protrude and harden. At will you are able to become a being of pure power: a carnal thing with a hunger for decimation. You have a form that you only refer to as 'Ravager'. Others know it as 'Monster'. Whatever title you wear one thing is certain: the world has not seen destruction as grand as you've made.";
     }
     if ((fPowerName == "TELEKINESIS" && mPowerName == "TELEKINESIS")) {
       powerName = "PSION";
-      Will = (int)(this.Will * 1.4f);
+      Will = this.Will * 1.4f;
       description = "Your parents were strong psychics in their own right, but you are even more than they could ever achieve. At a young age you could bend cars like hair in your fingers, growing up you could cause the clouds to stir on a whim. Your eyes see both the present and the future, a skill that neither of your parents possessed. You are an immensely powerful Psion. The world will bend at your feet whether they wish to or not.";
     }
     if ((fPowerName == "TELEKINESIS" && mPowerName == "PYROMANCY") || (fPowerName =="PYROMANCY" && mPowerName == "TELEKINESIS")) {
@@ -56,7 +57,7 @@ public class ChildPower {
     if (fPowerName == "PYROMANCY" && mPowerName == "PYROMANCY") {
       powerName = "FIRESTORM";
       description = "You are the sun, powerful and awesome. You are the light, radient and beautiful. You are the firestorm, consuming everything as you wish. The world burns with your steps. The flames will purify all. There can be nothing left in the wake of your fire.";
-      Fortitude = (int)(this.Fortitude * 1.4f);
+      Fortitude = this.Fortitude * 1.4f;
     }
     if ((fPowerName == "PYROMANCY" && mPowerName == "SHAPESHIFT") || fPowerName == "SHAPESHIFT" && mPowerName == "PYROMANCY") {
       powerName = "FLAME ASPECT";
@@ -65,17 +66,17 @@ public class ChildPower {
     if (fPowerName == "SHAPESHIFT" && mPowerName == "SHAPESHIFT") {
       powerName = "A THOUSAND FACES";
       description = "Your mastery over forms allows you to become anything you've ever dreamed of. It only takes one look for you to imitate someone completely.";
-      Agility = (int)(this.Agility * 1.4f);
+      Agility = this.Agility * 1.4f;
     }
 	}
 
-  float[] getGenerationalAttributes (ParentalPower father, ParentalPower mother) {
-    float cStrength = (((father.Strength + mother.Strength) /2) * 0.25f);
-    float cAgility = (((father.Agility + mother.Agility) /2) * 0.25f);
-    float cWill = (((father.Will + mother.Will) /2) * 0.25f);
-    float cFortitude = (((father.Fortitude + mother.Fortitude) /2) * 0.25f);
+  double[] getGenerationalAttributes (ParentalPower father, ParentalPower mother) {
+    double cStrength = (((father.Strength + mother.Strength) /2) * 0.25f);
+    double cAgility = (((father.Agility + mother.Agility) /2) * 0.25f);
+    double cWill = (((father.Will + mother.Will) /2) * 0.25f);
+    double cFortitude = (((father.Fortitude + mother.Fortitude) /2) * 0.25f);
 
-    float[] attributesArray = new float[] {
+    double[] attributesArray = new double[] {
       cStrength, cAgility, cWill, cFortitude
     };
